drivers/acrn: add the basic framework of ACRN char device driver
The ACRN hypervisor has a similar model to Xen. The Servic VM (like
Xen dom0) needs a middle layer that allow Service VM and DeviceModel
to communicate with ACRN hypervisor.

HSM driver is worked as this middle layer for ACRN hypervisor in
Service VM.

This patch adds the basic HSM driver framework and a char device with
name /dev/acrn_hsm is exported to Service VM user space to for user
space application(Devicemodel) could communicate with ACRN HSM module.

Signed-off-by: Liu Shuo <shuo.a.liu@intel.com>
Signed-off-by: Zhao Yakui <yakui.zhao@intel.com>
Cc: Shuo A Liu <shuo.a.liu@intel.com>
Cc: fengguang.wu@intel.com
Cc: Sean Christopherson <sean.j.christopherson@intel.com>
Cc: Dave Hansen <dave.hansen@intel.com>
Cc: Dan Williams <dan.j.williams@intel.com>