drivers/acrn: add service to obtain Power data transition
The px/cx data is critical to support the power transition. DM will get
these data to build DSDT for UOS. With this DSDT, UOS would have the
capability on power control if acpi-cpufreq/idle driver is enabled in
kernel.
Add the PM ioctl that is used to obtain the info of power state
so that the DM can construct the DSDT with Power frequence/C-state idle
for guest system.

v1-V2:  Add the __user tag prefix for copy_to_user

Co-developed-by: Jason Chen CJ <jason.cj.chen@intel.com>
Signed-off-by: Jason Chen CJ <jason.cj.chen@intel.com>
Co-developed-by: Victor Sun <victor.sun@intel.com>
Signed-off-by: Victor Sun <victor.sun@intel.com>
Signed-off-by: Zhao Yakui <yakui.zhao@intel.com>