drivers/acrn: Manage vcpu for ACRN char device
The vcpu is the basic execute env of running the guest OS. After it is
created, it is used by ACRN-hypervisor to run the guest OS. If SMP
is needed, multiple vcpus will be created. In order to start the
vcpu env with the specific configuration, the user-space device model is
allowed to set the register contexts for BSP vcpu.

As ACRN is mainly used for embedded IOT usage, no interface is provided
to destroy the vcpu explicitly. When the VM is destroyed, the low-level
ACRN hypervisor will free the corresponding vcpu implicitly.

Co-developed-by: Jason Chen CJ <jason.cj.chen@intel.com>
Signed-off-by: Jason Chen CJ <jason.cj.chen@intel.com>
Co-developed-by: Liu Shuo <shuo.a.liu@intel.com>
Signed-off-by: Liu Shuo <shuo.a.liu@intel.com>
Signed-off-by: Zhao Yakui <yakui.zhao@intel.com>