x86/acrn: add api to check whether the guest is privilege VM
The privilege VM is a special guest running on ACRN hypervisor.
The Devicemodel running in Privilege VM could manage other guests
by communicating with hypervisor.

The bit 0 of cpuid leaf 0x40000001 is extend by ACRN hypervisor
to detect whether the guest is privilege mode.

The patch 0006 which adds HSM driver will detect whether the
guest is privilege VM and load HSM driver when guest is.

Signed-off-by: Yin Fengwei <fengwei.yin@intel.com>
Cc: Shuo A Liu <shuo.a.liu@intel.com>
Cc: fengguang.wu@intel.com
Cc: Sean Christopherson <sean.j.christopherson@intel.com>
Cc: Dave Hansen <dave.hansen@intel.com>
Cc: Dan Williams <dan.j.williams@intel.com>
Cc: Jason Chen CJ <jason.cj.chen@intel.com>
Cc: Zhao Yakui <yakui.zhao@intel.com>