virt: acrn: Introduce I/O request management
For each VM, there is a shared 4-KByte memory region used for I/O
requests communication between the hypervisor and Service VM. An I/O
request is a 256-byte structure buffer, which is 'struct acrn_io_request',
that be filled by the hypervisor when a trapped I/O access happens in a
User VM.  ACRN userspace in the Service VM first allocates a 4-KByte
page and passes the GPA(Guest Physical Address) of the buffer to the
hypervisor. The buffer is used as an array of 16 I/O request slots with
each I/O request slot being 256 bytes. This array is indexed by vCPU ID.

An I/O client, which is 'struct acrn_ioreq_client', is responsible for
handling User VM I/O requests whose accessed GPA falls in a certain
range. Each VM has a list of registered I/O clients which are associated
with different I/O address ranges. There is a special client in each VM,
called default client, that handles all I/O requests that do not fit
into the range of any other I/O clients. In the current design the ACRN
userspace acts as the default I/O client for each User VM.

The state transitions of a ACRN I/O request are as follows.

   FREE -> PENDING -> PROCESSING -> COMPLETE -> FREE -> ...
                               \              /
                                +--> failed -+

FREE: this I/O request slot is empty
PENDING: a valid I/O request is pending in this slot
PROCESSING: the I/O request is being processed
COMPLETE: the I/O request has been processed

When a request is in COMPLETE or FREE state, the request is owned by the
hypervisor. HSM is in charge of processing the others.

The processing flow of I/O requests are listed as following:

a) The hypervisor makes an upcall, which is a notification interrupt, to
   the Service VM.
b) The upcall handler schedules a tasklet to dispatch I/O requests.
c) The tasklet looks for the PENDING I/O requests, assigns them to
   different registered clients based on the address of the I/O accesses,
   updates their state to PROCESSING, and notifies the
   corresponding client to handle.
d) The notified client handles the assigned I/O requests and updates
   their states to COMPLETE.
e) The HSM notifies the hypervisor of the completion via hypercalls.

Signed-off-by: Shuo A Liu <shuo.a.liu@intel.com>
Reviewed-by: Zhi Wang <zhi.a.wang@intel.com>
Cc: Zhi Wang <zhi.a.wang@intel.com>
Cc: Zhenyu Wang <zhenyuw@linux.intel.com>
Cc: Yu Wang <yu1.wang@intel.com>