virt: acrn: Add an interface for Service VM to control vCPU in partition mode
ACRN supports partition mode to achieve real-time requirements. In
partition mode, a CPU core can be dedicated to a vCPU of a User VM. The
local APIC of the dedicated CPU core can be passthrough to the User VM.
The Service VM controls the assignment of the CPU cores.

This patch introduces an interface for the Service VM to remove the
control of a CPU core from hypervisor perspective so that the CPU core
can be a dedicated CPU core of a User VM.

Signed-off-by: Shuo A Liu <shuo.a.liu@intel.com>
Reviewed-by: Zhi Wang <zhi.a.wang@intel.com>
Cc: Zhi Wang <zhi.a.wang@intel.com>
Cc: Zhenyu Wang <zhenyuw@linux.intel.com>
Cc: Yu Wang <yu1.wang@intel.com>